stripapplication/requirements.txt
cryptography==2.3.1
dj-database-url==0.5.0
Django==1.11
django-appconf==1.0.2
django-ipware==2.1.0
djangorestframework==3.6.3
enum34==1.1.6
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
jsonfield==2.0.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
stripe==2.8.1
urllib3==1.23
whitenoise==4.1
